backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.29
asyncpg==0.29.0
pydantic-settings==2.1.0
qdrant-client==1.9.0
openai==1.3.4
sentence-transformers==2.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6